Fix: Keyboard over fields for both platforms
