+++ src/components/TargetForm/styles.js
@@ -3,22 +3,35 @@ import { BLACK, BACKGROUND } from 'constants/colors';
 import typography from 'constants/typography';
 
 const styles = StyleSheet.create({
-  mainContainer: {
+  fullScreen: {
     position: 'absolute',
     top: 0,
     bottom: 0,
     left: 0,
     right: 0,
   },
-  container: {
+  mainContainer: {
     position: 'absolute',
+    backgroundColor: BACKGROUND,
+    bottom: 0,
+    left: 0,
+    right: 0,
+  },
+  androidContainer: {
+    position: 'absolute',
+    backgroundColor: BACKGROUND,
+    bottom: 0,
+    left: 0,
+    right: 0,
+    paddingHorizontal: 34,
+    paddingTop: 15,
+    paddingBottom: 26,
+  },
+  container: {
     backgroundColor: BACKGROUND,
     paddingHorizontal: 34,
     paddingTop: 15,
     paddingBottom: 26,
-    bottom: 0,
-    left: 0,
-    right: 0,
   },
   inputContainer: {
     marginBottom: 11,
